Implement AnnotationPolicy
The policy authorizes attaching labels to an annotation. Only users
who are editor or admin in one of the projects the annotation
belongs to are allowed to attach a label. And only labels belonging
to a label tree which is used by one of the projects the annotation
belongs to may be used.