Add links to used label trees
The links are always shown no matter if the user is allowed to
visit the label tree page or not since there is no straight forward
way to manage that. Admins can add any label tree that is available
for a project even if they have no permission to access it, so for
each newly attached label tree there has to be a separate check
whether the user can access the tree and whether to show a link or
not. This is currently not possible.

To mitigate the confusion of the user when they click on a link and
get a 403 response we must create meaningful error pages in
BiodataMiningGroup/dias-core#2.

Resolves #7