Fix annotation list scrolling behavior
If an annotation list item is selected it might be pushed down
by an unfolding label list item. Now the list adjusts the scrolling
so the selected annotation keeps the same relative position.

Also the automatic scrolling to show all selected annotations now
is only activated when annotations are selected in the image and
not in the sidebar list.

References #3