Remove unnecessary assignment of image transect
The original plan was to read the image dimensions from the image
files. This requires the full image file path with the URL of the
transect. However, the laserpoint detection always stores the
number of pixels of an image in the DB. So this information can be
used instead of reading the image and the transect URL is no longer
required.