Fix viewport center initialization
With the new extendChanged check, the viewport was always centered
for the first image that is displayed. But with the URL params the
viewport can have another position, too. Now the viewport is only
centered if the user switched from one image to the next (and the
extent changed).

References #24