Recompile JS
Sometimes the compiled JS is broken. I think this is because Gulp
can mix the ordering of the files. This can be a problem, e.g. with
vue/stores/map.js when the ZoomToNativeControl is not yet defined.
Maybe I have to think about another way to avoid this problem.