Always order annotations by ID in Largo
This will show the newest annotations first in Largo. We will use
this for the import of annotations proposed by the autoencoder
object detection. By importing the annotations from least salient
to most salient, the most salient proposals will be shown first in
Largo. Later we might implement a dedicated tool for this.