Fix positioning of annotation clips
If a vertical scroll bar was present, the position of the annotation
clips was skewed in relation to the position of the time indicator.
Absolute positioning (px) instead of relative positioning (%) fixes
this. Now annotation clips could overflow into the scroll bar but
this shouldn't happen very often and at least the display is correct.