Fix infinite loop when activating the lawnmower mode
The minimap component unintentionally made the whole Map object
reactive. The lawnmower mode modified the map view center which
depended on the map view extent. The map view extent however was
reactively updated when the center changed. This caused an infinite
loop. With the map object not being reactive, this is no issue.