Implement interpolatePoints() for the VideoAnnotation model
This method matches interpolatePoints() of the JS Annotation model
with the exception of line string and polygon interpolation. The
Polymorph (JS) interpolation is too complicated to implement. We
only need the interpolation on PHP side to be able to split
annotations. We simply won't support splitting of line strings
and polygons. Splitting is most useful for tracked annotations
(points and circles) anyway.

References #13