Fix edge case with annotation session
The edge case occurred if the annotation had multiple labels of
different users and the active annotation session was (later) set
to "hide other users annotations". When a user searched in Largo
for a label that was attached by another user but not themselves,
Largo would still return the annotation for that label (because
it has a label by the own user and was therefore visible in the
annotation session). Now the annotation is hidden in Largo unless
it has the label attached by the own user.