Add ffmpeg to app and worker images
This is done to support the biigle/videos module (and its tests)
out of the box. The app image already included the PHP SOAP extension
for biigle/label trees and the worker image has various Python packages.
It's easier to create a production instance or to develop BIIGLE if
all this is already included even though the images are a bit larger
because of this.