Fix bug with rounding of high-precision video annotation times
In the bug report the video annotation had a frame at a time like
200.12345. When video.currentTime was set to this value, it was
suddenly 200.12344 for some reason and the annotation was not shown
because the times did not match. Maybe the annotation was created
in a different browser (in my test it worked in Firefox but not in
Chrome). Now we round all times to a maximum of 4 decimals, which
should fix this issue.