Fix 502 Gateway Error if a PHP error occurs
The 502 was caused by the PHP error message passed on via SAPI logging to Nginx
but Nginx crashed with a "too big header" message.

This commit 1) disables SAPI logging and 2) configures Laravel to log to a file
instead of the errorlog. Any of these would fix the issue but I decided that
both should be done.

See: https://stackoverflow.com/a/68314973/1796523