Improve image cache size in annotation tool
The cache size is now controlled by the number of images that should
be preloaded.

The largest chunk of memory required is taken up by decoded images,
not the canvas elements of each cached image (see #653). Removing
the canvases had no great impact. But limiting the size of the cache
had, so the browser can know which images are "used" and which are
"unused" to better optimize the memory requirement.

Resolves #653