Fix annotation canvas immediate watchers
Immediate watchers are called before the created hook. But the modify
interactions is only created in the created hook so if canModify
did not change any more (which was not the case in the annotation tool
but *was* the case in the MAIA refine tabs), the modify interaction
was never activated.

Resolves https://github.com/biigle/maia/issues/124