Improve handling and configuration of GenerateFederatedSearchIndex job
If the default queue was congested, the job would not be run within
1 hour. When the "fetch index" request came from another instance,
it would attempt to generate the index synchronously because the
index was not there in the cache. This failed because generating the
index took longer than the 30 s request timeout.

Now the job is dispatched 30 min before the requests should arrive.
Also, the job can be pushed to a higher priority queue.