Fix video seeking on pause
This fixes a regression introduced in https://github.com/biigle/core/pull/726.
The video was supposed to seek to the current time on pause so
the displayed frame is the exact frame that matches currentTime
that is used for the annotations. However, on pause, the "dontSeek"
argument was set to an event object and not false, so the seeking
didn't happen.