Introduce query system that allows for combination of filters
This commit introduces a system that allows the selection of multiple values in the filter tab, so that it is now possible to combine multiple filters.
This comes with logical operators (is, is not, or, and). Changes the functioning of the API endpoints for filters of users and shapes as well.
In a further commit, tests will be introduced.