Make ApplyLargoSession job more robust against errors
If a job failed for some other reason than an exception during
saving, the largo_job_id was not cleaned up and no error was
shown. This is handled more robustly now. Also we need a mechanism
to prevent re-running of the failed job. Otherwise unexpected changes
could be applied. The exprected behavior is now that the user is
shown the error message and they try to save the session again.
The old job should be discarded then.