Align UI to Volume filter, fix remove filter bug, apply formatting
This commit applies changes to the UI for the Largo filtering that makes the UI more similar to the one in the Volume filters. This should make the UX better.
A bug is also solved that made the removal of filters not work correctly after removing the first filter.
Also, an automatic formatting is applied.