Fix issue with report CSV line count
There was a regression where putCsv() was replaced by put() but
the latter did not add a line break. This got noticed when an
abundance report was requested for a volume with a single image.
Since the volume title was not separated by the column titles with
a newline, the CSV ended up with only two rows. The Python script
assumed it to be empty.

See: https://github.com/biigle/core/commit/9f2f3957d302ccd4bc0e10bbf79b8ea26e4011e7