Change filter mechanism to allow information about selection and dismissal to be kept
This commit introduces a different mechanisms for the filtering of the annotations by creating a map of "allowed" annotations.
Before, the filtered annotations were kept in the annotationsCache, but that caused:
1. Duplication of data (rather than filtering the existing data, the new information about filters was newly added to the annotationsCache)
2. Information (as ex. about the dismissed annotations) got lost when adding or removing filters.
This commit  solves this problem by creating an array of allowed annotations and using that to filter the annotations already there.