Fix filters bug where the filters did not work correctly together
This commit fixes a bug where the usage of multiple filters with union did not work correctly: when using different filters (e.g. `user_id` and `shape_id`) with `union` set to 1 the different filters worked together as if they were exclusive.
A test is also added to each of the Api routes, so that this case is covered.