Fix bug that caused the possibility of adding filters values for filters it did not belong to
This commit fixes a bug that was caused by the fact that the `selectedFilterValue` variable did not reset when changing the `selectedFilter`.
Due to this, it was possible to add filter values from other filters, e.g. adding a filter from the users to shapes and vice versa.
We now reset the `selectedFilterValue` when moving to a different filter.