Fix bug that caused conflicting filters while using union to behave different than expected
This commit solves a problem where, if adding a `is` and `is not` filter for the same user/shape caused the union not to work correctly.
As an example, when adding in `union` mode filter `is Point` and `is not Point`, then we expect the filter to always return the entire annotation catalog.
If new filters are added in this case, this should not change anything, as they should act as inclusive filters.
New tests are added for this case.