Refactor filtering tab and largoContainer for filters/union management
This commit refactors the filtering tab and the largoContainer so that activeFilters and union values are not managed by the tab, rather they are now a prop passed by the tab to the tab.
The filtering tab now passes on the events by the annotationFilter to the largoContainer, which now manages it.
This should clarify the hierarchy of elements, reduce the number of repeated lines and allow the largoContainer the capacity of handling the filters directly in case of events.