Enable LabelBOT even if no annotations exist
This check made no sense, really, as LabelBOT would be activated
starting at the first annotation that the user creates. But also
with only 1 annotation, LabelBOT will likely not work well. We
need a more elaborate mechanism to show the reliability of LabelBOT
results. Until then it does not matter if 0 or 1 or 10 annotations
exist, so LabelBOT can be active in all cases.