Fix bug that caused weird behaviour with Favourites labelTree
This commit solves a nasty bug that basically caused weird behaviour in
the `Favourites` labelTree, e.g. being unable to remove from
favourites, the star not appearing and multiple favaourite assignment.
This was caused by the fact that the sortedTrees property was assigned
on `mounted`. This caused the `labelTree` component to be created after
the `labelTrees` component was created, meaning that the
`add-favourite` was emitted by the `labelTrees` component before the
`labelTree` component was created. TThe handler for the `add-favourite` event was not yet set, causing the weird behaviour.